+++ UnitTests/BraintreeCore_UnitTests/BTConfiguration_Tests.swift
@@ -159,18 +159,6 @@ class BTConfiguration_Tests: XCTestCase {
         XCTAssertEqual(configuration.applePaySupportedNetworks!, [PKPaymentNetwork.visa, PKPaymentNetwork.masterCard, PKPaymentNetwork.amex])
     }
 
-    func testApplePaySupportedNetworks_whenRunningBelowiOS9_doesNotReturnDiscover() {
-        let configurationJSON = BTJSON(value: [
-            "applePay": [ "supportedNetworks": ["discover"] ]
-            ])
-        let configuration = BTConfiguration(json: configurationJSON)
-
-        guard #available(iOS 9, *) else {
-            XCTAssertEqual(configuration.applePaySupportedNetworks!, [])
-            return
-        }
-    }
-
     @available(iOS 9.0, *)
     func testApplePaySupportedNetworks_whenSupportedNetworksIncludesDiscover_returnsSupportedNetworks() {
         let configurationJSON = BTJSON(value: [
@@ -179,6 +167,16 @@ class BTConfiguration_Tests: XCTestCase {
         let configuration = BTConfiguration(json: configurationJSON)
 
         XCTAssertEqual(configuration.applePaySupportedNetworks!, [PKPaymentNetwork.discover])
+    }
+
+    @available(iOS 12.0, *)
+    func testApplePaySupportedNetworks_whenSupportedNetworksIncludesMaestro_returnsSupportedNetworks() {
+        let configurationJSON = BTJSON(value: [
+            "applePay": [ "supportedNetworks": ["maestro"] ]
+            ])
+        let configuration = BTConfiguration(json: configurationJSON)
+
+        XCTAssertEqual(configuration.applePaySupportedNetworks!, [PKPaymentNetwork.maestro])
     }
 
     func testApplePaySupportedNetworks_doesNotPassesThroughUnknownValuesFromConfiguration() {